Add notebook
Add notebook for the analysis / visualization of manual OTS labels using nilearn. (Note: this will not run on the vanilla version of nilearn, because I specify view angle directly and not via a string, e.g. view = (210, 90) instead of view = 'ventral' . The latter option is the only possibility with the current official version of nilearn.)